Move file_blocks module into the iters module
