+++ src/dir_htree.rs
@@ -11,8 +11,8 @@ use crate::dir_entry::{DirEntry, DirEntryName};
 use crate::dir_entry_hash::dir_hash_md4_half;
 use crate::error::{Corrupt, Ext4Error, Incompatible};
 use crate::extent::{Extent, Extents};
-use crate::file_blocks::FileBlocks;
 use crate::inode::{Inode, InodeFlags, InodeIndex};
+use crate::iters::file_blocks::FileBlocks;
 use crate::path::PathBuf;
 use crate::util::{read_u16le, read_u32le, usize_from_u32};
 use crate::Ext4;
