+++ src/iters.rs
@@ -6,4 +6,5 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+pub(crate) mod file_blocks;
 pub(crate) mod read_dir;
