+++ src/iters/read_dir.rs
@@ -10,8 +10,8 @@ use crate::checksum::Checksum;
 use crate::dir_block::DirBlock;
 use crate::dir_entry::DirEntry;
 use crate::error::{Ext4Error, Incompatible};
-use crate::file_blocks::FileBlocks;
 use crate::inode::{Inode, InodeFlags, InodeIndex};
+use crate::iters::file_blocks::FileBlocks;
 use crate::path::PathBuf;
 use crate::util::usize_from_u32;
 use crate::Ext4;
