+++ src/lib.rs
@@ -110,7 +110,6 @@ mod dir_htree;
 mod error;
 mod extent;
 mod features;
-mod file_blocks;
 mod file_type;
 mod format;
 mod inode;
